backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
sqlalchemy==2.0.25
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0
PyPDF2==3.0.1
pytesseract==0.3.10
Pillow>=10.3.0
groq==0.4.2
python-dotenv==1.0.0
pytest==7.4.4
httpx==0.26.0
psycopg2-binary==2.9.11
faiss-cpu==1.9.0
sentence-transformers==3.3.1
tiktoken==0.8.0
